Add cron job + logger for updating new source requests

+++ src/updater/updater.service.ts
@@ -1,10 +1,13 @@
 import { Injectable } from '@nestjs/common'
 import { InjectRepository } from '@nestjs/typeorm'
 import { Repository } from 'typeorm'
+import * as delay from 'delay'
 import { UpdateRequest } from './update-request.entity'
 import { VkSource } from '@/data/entity/vk-source.entity'
 import { PostImportService } from '@/updater/post-import.service'
+import { CronLogger } from '@/logger/cron-logger.service'
 
+const API_DELAY = 1000
 const POST_COUNT = 100
 
 @Injectable()
@@ -15,6 +18,7 @@ export class Updater {
     @InjectRepository(VkSource)
     private readonly sourceRepository: Repository<VkSource>,
     private readonly postImportService: PostImportService,
+    private readonly logger: CronLogger,
   ) {}
 
   public async addRequest(sourceId: number, numPosts: number): Promise<void> {
@@ -24,11 +28,18 @@ export class Updater {
 
   public async update(): Promise<void> {
     const requests = await this.requestRepository.find({ take: 100 })
+    if (requests.length === 0) {
+      return
+    }
+
+    this.logger.info(`Updating ${requests.length} requests`)
 
     for (const request of requests) {
       await this.updateOne(request)
       await this.requestRepository.remove(request)
     }
+
+    this.logger.info(`Update finished`)
   }
 
   protected async updateOne(request: UpdateRequest): Promise<void> {
@@ -39,9 +50,15 @@ export class Updater {
     }
 
     let offset = source.parsedPosts
+    if (offset > 0) {
+      this.logger.debug(`Starting offset is ${offset}`)
+    }
     while (true) {
       // Nothing to do more
       if (source.isCompleted || source.parsedPosts >= request.numPosts) {
+        this.logger.debug(
+          `Parsed ${source.parsedPosts} from ${request.numPosts}`,
+        )
         return
       }
 
@@ -51,16 +68,19 @@ export class Updater {
         offset,
         POST_COUNT,
       )
+      this.logger.debug(
+        `Imported ${importedCount} posts from ${source.name} (${source.id})`,
+      )
 
       // Update source info
       source.parsedPosts += importedCount
       if (importedCount < POST_COUNT || importedCount === 0) {
         source.isCompleted = true
+        this.logger.debug(`Assume that we imported all posts from the source`)
       }
 
       await this.sourceRepository.save(source)
-
-      // Check if we got what we need
+      await delay(API_DELAY)
 
       offset += POST_COUNT
     }
